requirements.txt
requests
beautifulsoup4
html5lib
python-dotenv
google-genai